Fix endian detection across all build methods
Unfortunately Leptonica has been broken on big endian systems when
built with autotools because endianness.h was never used.

This commit ensures it is generated by Autotools, CMake, and the
static Makefile, so it has therefore been included unconditionally in
alltypes.h. This will break the native Visual Studio build but that
hasn't been maintained in a long time. For the static Makefile, the
detection should now work under any Make implementation, not just GNU.

With the detected endianness now preserved in a header, it is no
longer necessary to define L_LITTLE_ENDIAN or L_BIG_ENDIAN manually
when building against Leptonica. This is a substantial improvement as
forgetting to do so would have resulted in broken behaviour on big
endian systems despite a complete lack of errors or warnings. Note
that it will still respect your choice if you do decide to define
these manually.

Building universal binaries for OS X should theoretically work with
all build methods but this hasn't been tested. Feedback would be
appreciated.