Remove Autotools build of src/xtractprotos to fix #179
The allheaders target now triggers a build of xtractprotos in the prog
directory. I'm not sure if this is the proper way to do it but it's
good enough to have the following work as expected.

 ./configure
 cd src
 make allheaders

I have assumed that allheaders.h should be written to the source
directory rather than the build directory. Changing this would be
difficult.

It doesn't work when building outside of the source tree but this was
already broken. cpp, as invoked by xtractprotos, needs to locate
endianness.h and this is written to the build directory. Perhaps
xtractprotos should accept -I arguments or even pass all additional
arguments to cpp.