pnmio: Add support for P7 format (PAM)
The netpbm tools support a new "arbitrary" format.
See: http://netpbm.sourceforge.net/doc/pam.html

It is defined by the header "P7" followed a number of
lines with tags defining the attributes. The format
of the image data is always binary, as in P4, P5 and P6.

This format supports an optional alpha channel and
will be used automagically when a Pix with 32-bpp
and 4 samples per pixel is to be written by the
pixWriteStreamPnm() or pixWriteMemPnm() functions.

Not too many viewers do yet support this format.

For verifying and converting test images the netpbm tools
can be used. To create a sample 32 bpp rgba image do:
`pngtopam --alphapam test32-alpha.png >test32-alpha.pam`