Update README.md
At least 100 people have contributed to leptonica over the 16 years of its existence, but of these, there are six who have generously donated their time and talents to make major contributions to the library.  Without their dedication to this project, it is unlikely that leptonica would today have widespread use on multiple platforms.