Check for headers as well as libraries in configure
When I originally took on the autotools maintenance, I held the view
that it was not necessary to check for headers as the compiler would
fail eventually anyway. I've since decided that it's more polite to
check up front with a helpful error message, especially as it is
common for libraries to be installed without headers.