Fix Doxygen warnings.
After temporarily seting WARN_IF_UNDOCUMENTED = NO in Doxyfile,
the list of warnings and errors was not too big.

The most prevalent issue is the usage of < and > in comments,
because they are interpreted as HTML tag angle brackets.

In cases where replacing them with square brackets seems to
be acceptable, I did that. In other cases, like in formulas
where the notation `<x>` means average, they produce warnings.

The other changes were mostly missing </pre> tags or missing
\param or wrong named parameters.