fpix1: Fix warnings from lgtm
The previous fix did not satisfy lgtm because sizeof() results in a
size_t value which can have more than 32 bits.

Reordering the multiplication should be sufficient to fix this and
also avoids the type cast.

Change also nbytes here to be an unsigned value as size_t is unsigned, too.

Signed-off-by: Stefan Weil <sw@weilnetz.de>