Fixed leptonica pnm/pam reading of 16 bps rgb images.
People are now making 16 bps rgb pnm (type 6) files.
Leptonica's pnm read function was implemented before this was
allowed, and it makes a garbage image when reading from such
a file.  A tesseract user who makes these files was able to crash
tesseract with one of these images.

This has now been fixed.  There is, however, no standard for endian order
for 16 bps pnm/pam files.  Wikipedia says the de facto standard is
that the MSB is the first byte, but in the sample pnm file the image
data is in second byte (a surprise, because I assumed that was the LSB).