Fix uninitialized error (Tesseract #3089), found by Nam Le Duc.
This surfaced with valgrind on a library made with cmake, but not with
debug on autotools.  However, it can be seen with no-debug on autotools,
indicating that the issue arises from a malloc.
Made a program with a small image:
    pix1: small 1 bpp image
    pix2 = pixOpenBrick(NULL, pix1, 5, 5);
    pix3 = pixCloseSafeBrick(NULL, pix1, 4, 4);
    pix4 = pixSeedfillBinary(NULL, pix2, pix3, 4);
The problem is in pixCloseSafeBrick(), which adds a border before
doing the close and removes it afterward.  The functions that add
and remove the border must both initialize all pixels.

Also, use memset() instead of rasterops for pixSetAll() and pixClearAll().
This gets the pad bits as well as the image.