pnmio.c: avoid fseek()
The goal is to make it possible
to read a PNM/PAM image from stdin (netpbm-like),
which is currently not working: see issue #580.

The reason is the use of fseek(fp, -1, SEEK_CUR)
to "go back" after a fgetc(),
which doesn't work for stdin,
so it has been replaced through fscanf().