Fix rankhisto_reg test by fixing square root of negative number
The test fails in a debug build with compiler flag -ftrapv.
That compiler flag modifies the code to detect illegal
integer calculations.

An integer overflow was triggered because the difference of
two floating point values was negative (rounding error) and
the square root of negative values is undefined (NAN).

Converting NAN to integer results in an unexpected integer value.

Signed-off-by: Stefan Weil <sw@weilnetz.de>