Fix test for working gnuplot in reg_wrapper.sh
`gnuplot -e "set terminal png"` does not return an error if the
command fails, but prints an error message.

So instead of checking for an error the test now checks for a
non-empty output.

This fixes `make check` on hosts with a gnuplot which does not
support `set terminal png`.

Signed-off-by: Stefan Weil <sw@weilnetz.de>